[type-declaration] Rename NarrowTooWideReturnType to NarrowWideUnionReturnTypeRector
