--- rules-tests/DeadCode/Rector/FunctionLike/NarrowTooWideReturnTypeRector/Fixture/closure.php.inc
@@ -1,35 +1,0 @@
-<?php
-
-namespace Rector\Tests\DeadCode\Rector\FunctionLike\NarrowTooWideReturnTypeRector\Fixture;
-
-$callback = function(): string|int|array|null {
-    if (rand(0, 1)) {
-        return 'result';
-    }
-
-    if (rand(0, 1)) {
-        return 42;
-    }
-
-    return null;
-};
-
-?>
------
-<?php
-
-namespace Rector\Tests\DeadCode\Rector\FunctionLike\NarrowTooWideReturnTypeRector\Fixture;
-
-$callback = function(): string|int|null {
-    if (rand(0, 1)) {
-        return 'result';
-    }
-
-    if (rand(0, 1)) {
-        return 42;
-    }
-
-    return null;
-};
-
-?>
