--- rules-tests/DeadCode/Rector/FunctionLike/NarrowTooWideReturnTypeRector/Fixture/skip_docblock_call_on_arrow_function.php.inc
@@ -1,21 +1,0 @@
-<?php
-
-namespace Rector\Tests\DeadCode\Rector\FunctionLike\NarrowTooWideReturnTypeRector\Fixture;
-
-final class SkipDocblockCallOnArrowFunction
-{
-    public function run()
-    {
-        fn (): string|int => $this->execute();
-    }
-
-    /**
-     * @return string
-     */
-    private function execute()
-    {
-        return 1;
-    }
-}
-
-?>
