--- rules-tests/DeadCode/Rector/FunctionLike/NarrowTooWideReturnTypeRector/Fixture/skip_native_mixed_returns.php.inc
@@ -1,29 +1,0 @@
-<?php
-
-namespace Rector\Tests\DeadCode\Rector\FunctionLike\NarrowTooWideReturnTypeRector\Fixture;
-
-final class SkipFunctionLikesWithoutParameterTypes
-{
-    /**
-     * @param class-string<SomeInterface> $class
-     * @return class-string<SomeInterface>|int
-     */
-    public function bar($class): string|int
-    {
-        return $class;
-    }
-
-    /**
-     * @return class-string<\stdClass>|int
-     */
-    public function skipMixedByDoc(): string|int
-    {
-        /**
-         * @var class-string<\stdClass> $class
-         */
-        $class = get();
-        return $class;
-    }
-}
-
-?>
