--- rules-tests/DeadCode/Rector/FunctionLike/NarrowTooWideReturnTypeRector/Fixture/skip_non_final_classes.php.inc
@@ -1,18 +1,0 @@
-<?php
-
-namespace Rector\Tests\DeadCode\Rector\FunctionLike\NarrowTooWideReturnTypeRector\Fixture;
-
-class SkipNonFinalClass
-{
-    public function getData(): string|int|\DateTime
-    {
-        return 'hello';
-    }
-
-    protected function getInfo(): string|int|bool
-    {
-        return 'info';
-    }
-}
-
-?>
