--- rules-tests/DeadCode/Rector/FunctionLike/NarrowTooWideReturnTypeRector/Fixture/skip_return_all_types.php.inc
@@ -1,45 +1,0 @@
-<?php
-
-namespace Rector\Tests\DeadCode\Rector\FunctionLike\NarrowTooWideReturnTypeRector\Fixture;
-
-final class SkipReturnAllTypes
-{
-    public function getData(): string|int|float
-    {
-        if (rand(0, 1)) {
-            return 'text';
-        }
-
-        if (rand(0, 1)) {
-            return 0.42;
-        }
-
-        return 1000;
-    }
-
-    public function implicitReturn(): string|null
-    {
-        if (rand(0, 1)) {
-            return 'something';
-        }
-    }
-
-    function foo(): int|null
-    {
-        if (rand(0, 1)) {
-             return 1;
-        }
-
-        if (rand(0, 1)) {
-             return 2;
-        }
-
-        if (rand(0, 1)) {
-             return 3;
-        }
-
-        return null;
-    }
-}
-
-?>
