--- rules-tests/DeadCode/Rector/FunctionLike/NarrowTooWideReturnTypeRector/Fixture/skip_standalone_null.php.inc
@@ -1,18 +1,0 @@
-<?php
-
-namespace Rector\Tests\DeadCode\Rector\FunctionLike\NarrowTooWideReturnTypeRector\Fixture;
-
-final class SkipStandaloneNull
-{
-    public function foo(): ?string
-    {
-        return null;
-    }
-
-    public function foo(): string|null
-    {
-        return null;
-    }
-}
-
-?>
