--- rules-tests/DeadCode/Rector/FunctionLike/NarrowTooWideReturnTypeRector/Fixture/skip_true_false_to_become_bool.php.inc
@@ -1,16 +1,0 @@
-<?php
-
-namespace Rector\Tests\DeadCode\Rector\FunctionLike\NarrowTooWideReturnTypeRector\Fixture;
-
-final class SkipTrueFalseToBecomeBool
-{
-    public function run(): false|string
-    {
-        return false;
-    }
-
-    public function run(): true|string
-    {
-        return true;
-    }
-}
