--- rules-tests/DeadCode/Rector/FunctionLike/NarrowTooWideReturnTypeRector/Source/SomeAbstractClass.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Rector\Tests\DeadCode\Rector\FunctionLike\NarrowTooWideReturnTypeRector\Sourcet;
-
-abstract class SomeAbstractClass
-{
-    abstract public function process(): string|int|array;
-}
