--- rules-tests/DeadCode/Rector/FunctionLike/NarrowTooWideReturnTypeRector/Source/SomeInterface.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Rector\Tests\DeadCode\Rector\FunctionLike\NarrowTooWideReturnTypeRector\Source;
-
-interface SomeInterface
-{
-    public function getData(): string|int|bool;
-}
