--- rules-tests/DeadCode/Rector/FunctionLike/NarrowTooWideReturnTypeRector/config/configured_rule.php
@@ -1,12 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-use Rector\Config\RectorConfig;
-use Rector\DeadCode\Rector\FunctionLike\NarrowTooWideReturnTypeRector;
-use Rector\ValueObject\PhpVersionFeature;
-
-return static function (RectorConfig $rectorConfig): void {
-    $rectorConfig->rule(NarrowTooWideReturnTypeRector::class);
-    $rectorConfig->phpVersion(PhpVersionFeature::UNION_TYPES);
-};
