+++ rules-tests/DeadCode/Rector/FunctionLike/NarrowWideUnionReturnTypeRector/Fixture/skip_abstracts.php.inc
@@ -1,0 +1,28 @@
+<?php
+
+namespace Rector\Tests\DeadCode\Rector\FunctionLike\NarrowWideUnionReturnTypeRector\Fixture;
+
+abstract class SkipAbstractClass
+{
+    abstract public function process(): string|int|array;
+
+    public function getData(): string|int|bool
+    {
+        return 'data';
+    }
+}
+
+interface SkipInterface
+{
+    public function handle(): string|int|null;
+}
+
+trait SkipTrait
+{
+    public function process(): string|array|null
+    {
+        return 'processed';
+    }
+}
+
+?>
