+++ rules-tests/DeadCode/Rector/FunctionLike/NarrowWideUnionReturnTypeRector/Fixture/skip_docblock_call_on_arrow_function.php.inc
@@ -1,0 +1,21 @@
+<?php
+
+namespace Rector\Tests\DeadCode\Rector\FunctionLike\NarrowWideUnionReturnTypeRector\Fixture;
+
+final class SkipDocblockCallOnArrowFunction
+{
+    public function run()
+    {
+        fn (): string|int => $this->execute();
+    }
+
+    /**
+     * @return string
+     */
+    private function execute()
+    {
+        return 1;
+    }
+}
+
+?>
