+++ rules-tests/DeadCode/Rector/FunctionLike/NarrowWideUnionReturnTypeRector/Fixture/skip_non_final_classes.php.inc
@@ -1,0 +1,18 @@
+<?php
+
+namespace Rector\Tests\DeadCode\Rector\FunctionLike\NarrowWideUnionReturnTypeRector\Fixture;
+
+class SkipNonFinalClass
+{
+    public function getData(): string|int|\DateTime
+    {
+        return 'hello';
+    }
+
+    protected function getInfo(): string|int|bool
+    {
+        return 'info';
+    }
+}
+
+?>
