+++ rules-tests/DeadCode/Rector/FunctionLike/NarrowWideUnionReturnTypeRector/Fixture/skip_parent_and_child_combine.php.inc
@@ -1,0 +1,13 @@
+<?php
+
+namespace Rector\Tests\DeadCode\Rector\FunctionLike\NarrowWideUnionReturnTypeRector\Fixture;
+
+final class SkipParentAndChildCombine
+{
+    public function run(): \DateTimeInterface|\DateTimeImmutable
+    {
+        return rand(0, 1)
+            ? new \DateTime('now')
+            : new \DateTimeImmutable('now');
+    }
+}
