+++ rules-tests/DeadCode/Rector/FunctionLike/NarrowWideUnionReturnTypeRector/Fixture/skip_standalone_null.php.inc
@@ -1,0 +1,18 @@
+<?php
+
+namespace Rector\Tests\DeadCode\Rector\FunctionLike\NarrowWideUnionReturnTypeRector\Fixture;
+
+final class SkipStandaloneNull
+{
+    public function foo(): ?string
+    {
+        return null;
+    }
+
+    public function foo(): string|null
+    {
+        return null;
+    }
+}
+
+?>
