+++ rules-tests/DeadCode/Rector/FunctionLike/NarrowWideUnionReturnTypeRector/Fixture/skip_true_false_to_become_bool.php.inc
@@ -1,0 +1,16 @@
+<?php
+
+namespace Rector\Tests\DeadCode\Rector\FunctionLike\NarrowWideUnionReturnTypeRector\Fixture;
+
+final class SkipTrueFalseToBecomeBool
+{
+    public function run(): false|string
+    {
+        return false;
+    }
+
+    public function run(): true|string
+    {
+        return true;
+    }
+}
