+++ rules-tests/DeadCode/Rector/FunctionLike/NarrowWideUnionReturnTypeRector/Source/SomeAbstractClass.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Rector\Tests\DeadCode\Rector\FunctionLike\NarrowWideUnionReturnTypeRector\Sourcet;
+
+abstract class SomeAbstractClass
+{
+    abstract public function process(): string|int|array;
+}
