+++ rules-tests/DeadCode/Rector/FunctionLike/NarrowWideUnionReturnTypeRector/Source/SomeInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Rector\Tests\DeadCode\Rector\FunctionLike\NarrowWideUnionReturnTypeRector\Source;
+
+interface SomeInterface
+{
+    public function getData(): string|int|bool;
+}
