+++ rules/CodingStyle/ValueObject/ConditionAndResult.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace Rector\CodingStyle\ValueObject;
 
-use PhpParser\Node\Expr\Variable;
 use PhpParser\Node\Expr;
 use PhpParser\Node\Expr\BinaryOp\Identical;
+use PhpParser\Node\Expr\Variable;
 use Webmozart\Assert\Assert;
 
 final readonly class ConditionAndResult
