+++ src/Config/Level/DeadCodeLevel.php
@@ -36,7 +36,7 @@ use Rector\DeadCode\Rector\For_\RemoveDeadIfForeachForRector;
 use Rector\DeadCode\Rector\For_\RemoveDeadLoopRector;
 use Rector\DeadCode\Rector\Foreach_\RemoveUnusedForeachKeyRector;
 use Rector\DeadCode\Rector\FuncCall\RemoveFilterVarOnExactTypeRector;
-use Rector\DeadCode\Rector\FunctionLike\NarrowTooWideReturnTypeRector;
+use Rector\DeadCode\Rector\FunctionLike\NarrowWideUnionReturnTypeRector;
 use Rector\DeadCode\Rector\FunctionLike\RemoveDeadReturnRector;
 use Rector\DeadCode\Rector\If_\ReduceAlwaysFalseIfOrRector;
 use Rector\DeadCode\Rector\If_\RemoveAlwaysTrueIfConditionRector;
@@ -148,6 +148,6 @@ final class DeadCodeLevel
 
         RemoveArgumentFromDefaultParentCallRector::class,
         RemoveNullArgOnNullDefaultParamRector::class,
-        NarrowTooWideReturnTypeRector::class,
+        NarrowWideUnionReturnTypeRector::class,
     ];
 }
